+++ crates/iota-indexer/src/ingestion/primary/mod.rs
@@ -1,5 +1,6 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+pub(crate) mod orchestration;
 pub(crate) mod persist;
 pub(crate) mod prepare;
